add internal/sign library packages for signer job
This commit adds library support for creating and managing k8s jobs that
will back the Sign API. It addsi the internal/sigfn directory structure
including three types
signer - creates the job tmplate
signJobManager - creates, manages, and reports the completion of the job
helper - caclulates the effective Sign configuration from the
kerrnelmapping and defaults.

This closely follows the exisitng internal/build (kaniko)
implemenataion, mainly because it essentially deos the same thing (run a
job external to the controller)

This commit only adds libraries, it does not add controller support.
km.Sign will still not work after this commit.
This is this code can be suitably reviewed and ammended as needed
without requiring extensive rebasinig to keep it current.i Controlelr
support is the final piece needed and will be added in a seperate commit.